requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
autopep8==1.6.0
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
discord.py==2.0.0a4391+gc61e407b
gevent==22.10.2
gevent-eventemitter==2.1
greenlet==2.0.2
idna==3.3
multidict==6.0.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pkg_resources==0.0.0
platformdirs==3.2.0
protobuf==3.20.3
pycodestyle==2.8.0
pycryptodomex==3.17
python-dotenv==0.20.0
requests==2.28.2
six==1.16.0
steam==1.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.15
vdf==3.4
yarl==1.7.2
zope.event==4.6
zope.interface==6.0
